added logCreateArticle to action loader and moved it to insight-analytics-actions folder
