+++ packages/headless/src/features/analytics/insight-analytics-actions-loader.ts
@@ -42,6 +42,8 @@ import {InsightAction} from './analytics-utils';
 import {
   logInsightInterfaceChange,
   logInsightInterfaceLoad,
+  logInsightCreateArticle,
+  CreateArticleMetadata,
 } from './insight-analytics-actions';
 
 export type {
@@ -55,6 +57,7 @@ export type {
   LogFacetSelectActionCreatorPayload,
   SmartSnippetFeedback,
   StaticFilterValueMetadata,
+  CreateArticleMetadata,
 };
 
 /**
@@ -81,6 +84,14 @@ export interface InsightAnalyticsActionCreators {
    * @returns A dispatchable action.
    */
   logInterfaceChange(): InsightAction;
+
+  /**
+   * The event to log when the create article button is clicked.
+   *
+   * @param payload - The metadata of tied to the create article event.
+   * @returns A dispatchable action.
+   */
+  logCreateArticle(payload: CreateArticleMetadata): InsightAction;
 
   /**
    * The event to log when a category facet breadcrumb is deselected.
@@ -252,6 +263,7 @@ export function loadInsightAnalyticsActions(
     logClearBreadcrumbs,
     logInterfaceLoad: logInsightInterfaceLoad,
     logInterfaceChange: logInsightInterfaceChange,
+    logCreateArticle: logInsightCreateArticle,
     logCategoryFacetBreadcrumb,
     logFacetBreadcrumb,
     logFacetClearAll,
