+++ packages/headless/src/features/analytics/insight-analytics-actions.test.ts
@@ -3,24 +3,31 @@ import {buildMockInsightEngine} from '../../test/mock-engine';
 import {buildMockInsightState} from '../../test/mock-insight-state';
 import {getConfigurationInitialState} from '../configuration/configuration-state';
 import {
+  logInsightCreateArticle,
   logInsightInterfaceChange,
   logInsightInterfaceLoad,
 } from './insight-analytics-actions';
 
 const mockLogInterfaceLoad = jest.fn();
 const mockLogInterfaceChange = jest.fn();
+const mockLogCreateArticle = jest.fn();
 
 const exampleSubject = 'example subject';
 const exampleDescription = 'example description';
 const exampleCaseId = '1234';
 const exampleCaseNumber = '5678';
 const exampleOriginLevel2 = 'exampleOriginLevel2';
+const exampleCreateArticleMetadata = {
+  articleType: 'Knowledge__kav',
+  triggeredBy: 'CreateArticleButton',
+};
 
 jest.mock('coveo.analytics', () => {
   const mockCoveoInsightClient = jest.fn(() => ({
     disable: jest.fn(),
     logInterfaceLoad: mockLogInterfaceLoad,
     logInterfaceChange: mockLogInterfaceChange,
+    logCreateArticle: mockLogCreateArticle,
   }));
 
   return {
@@ -98,6 +105,42 @@ describe('logInterfaceChange', () => {
     expect(mockLogInterfaceChange).toBeCalledTimes(1);
     expect(mockLogInterfaceChange.mock.calls[0][0]).toStrictEqual(
       expectedPayload
+    );
+  });
+});
+
+describe('the analytics related to the create article feature in the insight use case', () => {
+  const engine = buildMockInsightEngine({
+    state: buildMockInsightState({
+      insightCaseContext: {
+        caseContext: {
+          Case_Subject: exampleSubject,
+          Case_Description: exampleDescription,
+        },
+        caseId: exampleCaseId,
+        caseNumber: exampleCaseNumber,
+      },
+    }),
+  });
+
+  it('should log #logCreateArticle with the right payload', async () => {
+    await engine.dispatch(
+      logInsightCreateArticle(exampleCreateArticleMetadata)
+    );
+    const expectedCaseContext = {
+      caseContext: {
+        Case_Subject: exampleSubject,
+        Case_Description: exampleDescription,
+      },
+      caseId: exampleCaseId,
+      caseNumber: exampleCaseNumber,
+    };
+    expect(mockLogCreateArticle).toHaveBeenCalledTimes(1);
+    expect(mockLogCreateArticle.mock.calls[0][0]).toStrictEqual(
+      exampleCreateArticleMetadata
+    );
+    expect(mockLogCreateArticle.mock.calls[0][1]).toStrictEqual(
+      expectedCaseContext
     );
   });
 });
