+++ packages/headless/src/features/analytics/insight-analytics-actions.ts
@@ -1,6 +1,11 @@
 import {getCaseContextAnalyticsMetadata} from '../case-context/case-context-state';
 import {InsightAction, makeInsightAnalyticsAction} from './analytics-utils';
 
+export interface CreateArticleMetadata {
+  articleType: string;
+  triggeredBy: string;
+}
+
 export const logInsightInterfaceLoad = (): InsightAction =>
   makeInsightAnalyticsAction('analytics/interface/load', (client, state) =>
     client.logInterfaceLoad(
@@ -14,4 +19,14 @@ export const logInsightInterfaceChange = (): InsightAction =>
       ...getCaseContextAnalyticsMetadata(state.insightCaseContext),
       interfaceChangeTo: state.configuration.analytics.originLevel2,
     });
+  });
+
+export const logInsightCreateArticle = (
+  createArticleMetadata: CreateArticleMetadata
+): InsightAction =>
+  makeInsightAnalyticsAction('analytics/createArticle', (client, state) => {
+    return client.logCreateArticle(
+      createArticleMetadata,
+      getCaseContextAnalyticsMetadata(state.insightCaseContext)
+    );
   });
